Issue 695: Refactored domain package: created packages: virtualmachine,hardware,software,network
