+++ sandbox-providers/tmrk-enterprisecloud/src/main/java/org/jclouds/tmrk/enterprisecloud/domain/network/Nics.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.jclouds.tmrk.enterprisecloud.domain;
+package org.jclouds.tmrk.enterprisecloud.domain.network;
 
 import com.google.common.collect.Sets;
 
@@ -25,33 +25,44 @@ import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
 /**
- * Container for DeviceNetwork items
- * <xs:complexType name="DeviceNetworks">
+ * Wraps individual VirtualNic elements.
+ * Needed because parsing is done with JAXB and it does not handle Generic collections
+ * <xs:complexType name="Nics">
  * @author Jason King
  */
-public class DeviceNetworks {
+public class Nics {
 
-    private LinkedHashSet<DeviceNetwork> deviceNetworks = Sets.newLinkedHashSet();
+    private LinkedHashSet<VirtualNic> nics = Sets.newLinkedHashSet();
 
-    protected DeviceNetworks() {
-        //For JAXB
+    @XmlElement(name = "Nic")
+    public void setVirtualNic(VirtualNic nic) {
+        this.nics.add(nic);
     }
 
-    @XmlElement(name = "Network")
-    void setDeviceNetwork(DeviceNetwork deviceNetwork) {
-        checkNotNull(deviceNetwork,"deviceNetwork");
-        this.deviceNetworks.add(deviceNetwork);
+    public Set<VirtualNic> getVirtualNics() {
+        return Collections.unmodifiableSet(nics);
     }
 
-    public Set<DeviceNetwork> getDeviceNetworks() {
-        return Collections.unmodifiableSet(deviceNetworks);
-    }
-    
     @Override
-    public String toString() {
-        return "[deviceNetworks="+deviceNetworks+"]";
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        Nics nics1 = (Nics) o;
+
+        if (!nics.equals(nics1.nics)) return false;
+
+        return true;
     }
+
+    @Override
+    public int hashCode() {
+        return nics.hashCode();
+    }
+
+    public String toString() {
+        return "["+ nics.toString()+"]";
+    }
+
 }
