+++ sandbox-providers/tmrk-enterprisecloud/src/main/java/org/jclouds/tmrk/enterprisecloud/features/VirtualMachineAsyncClient.java
@@ -22,9 +22,9 @@ import com.google.common.util.concurrent.ListenableFuture;
 import org.jclouds.http.filters.BasicAuthentication;
 import org.jclouds.rest.annotations.*;
 import org.jclouds.rest.functions.ReturnNullOnNotFoundOr404;
-import org.jclouds.tmrk.enterprisecloud.domain.AssignedIpAddresses;
-import org.jclouds.tmrk.enterprisecloud.domain.VirtualMachine;
-import org.jclouds.tmrk.enterprisecloud.domain.VirtualMachines;
+import org.jclouds.tmrk.enterprisecloud.domain.network.AssignedIpAddresses;
+import org.jclouds.tmrk.enterprisecloud.domain.vm.VirtualMachine;
+import org.jclouds.tmrk.enterprisecloud.domain.vm.VirtualMachines;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
