+++ sandbox-providers/tmrk-enterprisecloud/src/test/java/org/jclouds/tmrk/enterprisecloud/xml/AssignedIpAddressesJAXBParsingTest.java
@@ -35,8 +35,8 @@ import org.jclouds.rest.AuthorizationException;
 import org.jclouds.rest.BaseRestClientTest;
 import org.jclouds.rest.RestContextSpec;
 import org.jclouds.rest.internal.RestAnnotationProcessor;
-import org.jclouds.tmrk.enterprisecloud.domain.*;
-import org.jclouds.tmrk.enterprisecloud.domain.VirtualMachine.VirtualMachineStatus;
+import org.jclouds.tmrk.enterprisecloud.domain.network.AssignedIpAddresses;
+import org.jclouds.tmrk.enterprisecloud.domain.network.DeviceNetwork;
 import org.jclouds.tmrk.enterprisecloud.features.VirtualMachineAsyncClient;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
