+++ sandbox-providers/tmrk-enterprisecloud/src/test/java/org/jclouds/tmrk/enterprisecloud/xml/VirtualMachineJAXBParsingTest.java
@@ -36,7 +36,18 @@ import org.jclouds.rest.BaseRestClientTest;
 import org.jclouds.rest.RestContextSpec;
 import org.jclouds.rest.internal.RestAnnotationProcessor;
 import org.jclouds.tmrk.enterprisecloud.domain.*;
-import org.jclouds.tmrk.enterprisecloud.domain.VirtualMachine.VirtualMachineStatus;
+import org.jclouds.tmrk.enterprisecloud.domain.hardware.HardwareConfiguration;
+import org.jclouds.tmrk.enterprisecloud.domain.hardware.Memory;
+import org.jclouds.tmrk.enterprisecloud.domain.hardware.VirtualDisk;
+import org.jclouds.tmrk.enterprisecloud.domain.network.AssignedIpAddresses;
+import org.jclouds.tmrk.enterprisecloud.domain.network.DeviceNetwork;
+import org.jclouds.tmrk.enterprisecloud.domain.network.NetworkReference;
+import org.jclouds.tmrk.enterprisecloud.domain.network.VirtualNic;
+import org.jclouds.tmrk.enterprisecloud.domain.software.OperatingSystem;
+import org.jclouds.tmrk.enterprisecloud.domain.software.ToolsStatus;
+import org.jclouds.tmrk.enterprisecloud.domain.vm.VirtualMachine.VirtualMachineStatus;
+import org.jclouds.tmrk.enterprisecloud.domain.vm.VirtualMachine;
+import org.jclouds.tmrk.enterprisecloud.domain.vm.VirtualMachineIpAddresses;
 import org.jclouds.tmrk.enterprisecloud.features.VirtualMachineAsyncClient;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
